requirements.txt
streamlit
python-dotenv
plotly
PyGithub
pydantic
pyarrow

# Compatibilidade cruzada Python 3.12 e 3.13
# - Em Python <3.13, usamos NumPy 1.26.x e pandas 2.2.x
# - Em Python >=3.13, usamos NumPy 2.x e pandas 2.3+
numpy==1.26.4; python_version < "3.13"
numpy>=2.1,<3; python_version >= "3.13"
pandas>=2.2,<2.3; python_version < "3.13"
pandas>=2.3; python_version >= "3.13"
